# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AttachTagRequest:
    properties:
      tag_id:
        type: integer
    required:
    - tag_id
    type: object
...
  dto.CreateColumnRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  dto.CreateTagRequest:
    properties:
      color:
        maxLength: 50
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  dto.TagResponse:
    properties:
      color:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.UpdateColumnRequest:
    properties:
      auto_pin:
        type: boolean
      name:
        type: string
      position:
        type: number
    type: object
  dto.UpdateTagRequest:
    properties:
      color:
        maxLength: 50
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
  model.Card:
    properties:
      checklists:
        items:
          $ref: '#/definitions/model.Checklist'
        type: array
      column_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_pinned:
        type: boolean
      position:
        type: number
      start_time:
        type: string
      story_point:
        type: integer
      tags:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.Checklist:
    properties:
      card_id:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.ChecklistItem'
        type: array
      position:
        type: number
      title:
        type: string
    type: object
  model.ChecklistItem:
    properties:
      checklist_id:
        type: integer
      completed:
        type: boolean
      id:
        type: integer
      position:
        type: number
      text:
        type: string
    type: object
  model.Column:
    properties:
      auto_pin:
        type: boolean
      board_id:
        type: integer
      cards:
        items:
          $ref: '#/definitions/model.Card'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      position:
        type: number
      updated_at:
        type: string
    type: object
  model.Tag:
    properties:
      color:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:34115
info:
  contact: {}
  description: Kanban board with pin mode API
  title: Pinflow API
  version: "1.0"
paths:
...
  /boards/{id}/columns:
    post:
      consumes:
      - application/json
      parameters:
      - description: Board ID
        in: path
        name: id
        required: true
        type: integer
      - description: Column data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateColumnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Column'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a column in a board
      tags:
      - columns
...
  /cards/{id}/tags:
    post:
      consumes:
      - application/json
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AttachTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TagResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Attach a tag to a card
      tags:
      - tags
  /cards/{id}/tags/{tagId}:
    delete:
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag ID
        in: path
        name: tagId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Detach a tag from a card
      tags:
      - tags
...
  /columns/{id}:
    delete:
      parameters:
      - description: Column ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a column
      tags:
      - columns
    patch:
      consumes:
      - application/json
      parameters:
      - description: Column ID
        in: path
        name: id
        required: true
        type: integer
      - description: Column update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateColumnRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Column'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a column (rename, auto_pin, position)
      tags:
      - columns
...
  /tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TagResponse'
            type: array
      summary: List all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      parameters:
      - description: Tag data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TagResponse'
      summary: Create or get a tag by name
      tags:
      - tags
  /tags/{id}:
    delete:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a tag and its card associations
      tags:
      - tags
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a tag's name and/or color
      tags:
      - tags
swagger: "2.0"